refactor history.ts file in utils
