+++ bridge-ui/src/stores/historyStore.ts
@@ -7,29 +7,34 @@ import { BridgeTransaction } from "@/types";
 
 export type HistoryState = {
   isLoading: boolean;
-  history: Record<
-    string,
-    { transactions: BridgeTransaction[]; lastL1FetchedBlockNumber: bigint; lastL2FetchedBlockNumber: bigint }
-  >;
+  // history: Record<
+  //   string,
+  //   { transactions: BridgeTransaction[]; lastL1FetchedBlockNumber: bigint; lastL2FetchedBlockNumber: bigint }
+  // >;
+  completeTxHistory: Record<string, BridgeTransaction>;
 };
 
 export type HistoryActions = {
   setIsLoading: (isLoading: boolean) => void;
-  setTransactions: (
-    key: string,
-    transactions: BridgeTransaction[],
-    lastL1FetchedBlockNumber?: bigint,
-    lastL2FetchedBlockNumber?: bigint,
-  ) => void;
-  getTransactionsByKey: (key: string) => BridgeTransaction[];
-  getFromBlockNumbers: (key: string) => { l1FromBlock: bigint; l2FromBlock: bigint };
+  // setTransactions: (
+  //   key: string,
+  //   transactions: BridgeTransaction[],
+  //   lastL1FetchedBlockNumber?: bigint,
+  //   lastL2FetchedBlockNumber?: bigint,
+  // ) => void;
+  setCompleteTx: (key: string, transaction: BridgeTransaction) => void;
+  // getTransactionsByKey: (key: string) => BridgeTransaction[];
+  // getFromBlockNumbers: (key: string) => { l1FromBlock: bigint; l2FromBlock: bigint };
+  getCompleteTx: (key: string) => BridgeTransaction | undefined;
+  getKey: (transaction: BridgeTransaction) => string;
 };
 
 export type HistoryStore = HistoryState & HistoryActions;
 
 export const defaultInitState: HistoryState = {
-  history: {},
   isLoading: false,
+  // history: {},
+  completeTxHistory: {},
 };
 
 export const useHistoryStore = createWithEqualityFn<HistoryStore>()(
@@ -37,36 +42,50 @@ export const useHistoryStore = createWithEqualityFn<HistoryStore>()(
     (set, get) => ({
       ...defaultInitState,
       setIsLoading: (isLoading) => set({ isLoading }),
-      setTransactions: (key, transactions, lastL1FetchedBlockNumber, lastL2FetchedBlockNumber) =>
+      // setTransactions: (key, transactions, lastL1FetchedBlockNumber, lastL2FetchedBlockNumber) =>
+      //   set((state) => ({
+      //     history: {
+      //       ...state.history,
+      //       [key]: {
+      //         transactions,
+      //         lastL1FetchedBlockNumber: lastL1FetchedBlockNumber || 0n,
+      //         lastL2FetchedBlockNumber: lastL2FetchedBlockNumber || 0n,
+      //       },
+      //     },
+      //   })),
+      setCompleteTx: (key, transaction) =>
         set((state) => ({
-          history: {
-            ...state.history,
-            [key]: {
-              transactions,
-              lastL1FetchedBlockNumber: lastL1FetchedBlockNumber || 0n,
-              lastL2FetchedBlockNumber: lastL2FetchedBlockNumber || 0n,
-            },
+          completeTxHistory: {
+            ...state.completeTxHistory,
+            [key]: transaction,
           },
         })),
-      getTransactionsByKey: (key) => {
-        const { history } = get();
-        return history[key]?.transactions ?? [];
+      getCompleteTx: (key) => {
+        const { completeTxHistory } = get();
+        return completeTxHistory[key];
       },
-      getFromBlockNumbers: (key) => {
-        const { history } = get();
-
-        if (isEmptyObject(history) || !history[key]) {
-          return {
-            l1FromBlock: 0n,
-            l2FromBlock: 0n,
-          };
-        }
-
-        return {
-          l1FromBlock: history[key].lastL1FetchedBlockNumber,
-          l2FromBlock: history[key].lastL2FetchedBlockNumber,
-        };
+      getKey: (transaction) => {
+        return "";
       },
+      // getTransactionsByKey: (key) => {
+      //   const { history } = get();
+      //   return history[key]?.transactions ?? [];
+      // },
+      // getFromBlockNumbers: (key) => {
+      //   const { history } = get();
+
+      //   if (isEmptyObject(history) || !history[key]) {
+      //     return {
+      //       l1FromBlock: 0n,
+      //       l2FromBlock: 0n,
+      //     };
+      //   }
+
+      //   return {
+      //     l1FromBlock: history[key].lastL1FetchedBlockNumber,
+      //     l2FromBlock: history[key].lastL2FetchedBlockNumber,
+      //   };
+      // },
     }),
     {
       name: "history-storage",
