+++ bridge-ui/src/utils/history/getBridgeTransactionKey.ts
@@ -1,0 +1,5 @@
+import { BridgeTransaction } from "@/types";
+
+export function getBridgeTransactionKey(transaction: BridgeTransaction): string {
+  return `${transaction.bridgingTx}-${transaction.claimingTx}`;
+}
