+++ bridge-ui/src/utils/history/index.ts
@@ -1,0 +1,2 @@
+export { fetchTransactionsHistory } from "./fetchTransactionsHistory";
+export { getBridgeTransactionKey } from "./getBridgeTransactionKey";
